requirements.txt
streamlit==1.46.1
sentence-transformers==2.2.2
faiss-cpu==1.7.4
numpy>=1.26.0
pandas==2.0.3
scikit-learn==1.3.0
huggingface_hub==0.10.1
transformers==4.24.0